[Linux] compilation fix for PR #4099[ci skip]

+++ cocos/2d/CCDirector.cpp
@@ -831,7 +831,7 @@ void Director::calculateMPF()
 }
 
 // returns the FPS image data pointer and len
-void Director::getFPSImageData(unsigned char** datapointer, unsigned int* length)
+void Director::getFPSImageData(unsigned char** datapointer, long* length)
 {
     // XXX fixed me if it should be used 
     *datapointer = cc_fps_images_png;
@@ -855,7 +855,7 @@ void Director::createStatsLabel()
     Texture2D::PixelFormat currentFormat = Texture2D::getDefaultAlphaPixelFormat();
     Texture2D::setDefaultAlphaPixelFormat(Texture2D::PixelFormat::RGBA4444);
     unsigned char *data = nullptr;
-    unsigned int dataLength = 0;
+    long dataLength = 0;
     getFPSImageData(&data, &dataLength);
 
     Image* image = new Image();
